New number treatment to deal with constants.
Constants like 1 or 0 can be either Integer_8, or Integer_64, etc.
To deal with this, we have a new type that represents any Integer,
which can then be casted to Integer_8 or Integer_64. Internally
the "any integer" is represented by int64, then casted appropriately.

+++ compiler/tests/lib/externals.c
@@ -17,7 +17,7 @@ pointer_8_put (char* p, int i, int8_t v)
 
 struct string
 {
-  int length;
+  int64_t length;
   char* data;
 };
